gate tests and utils that require live_visualization
