+++ rust/foxglove/src/tests.rs
@@ -1,4 +1,5 @@
 use std::{error::Error, fmt::Display};
+#[cfg(feature = "live_visualization")]
 mod logging;
 mod schemas;
